test for batch verify
